\name{accessors.forest}
\alias{[,forest-method}
\alias{accessors.forest}
\title{Accessor method for forest class}
\description{
  Methods to access public slots in the forest class.
}
\usage{
\S4method{[}{forest}(x, i)
}
\arguments{
  \item{x}{Object of class inheriting from \code{"\link{forest-class}"}.}
  \item{i}{String value that is either "n", "p", "min_node_obs", "max_depth", "numsamps", "numvars", "numboots", "numnodes", "model", "x", "y", or "fmla"}
}
