\name{easy_2var_data}
\docType{data}
\alias{easy_2var_data}
\title{Easy to fit dataset for decision trees}
\description{Fake data that should be easy for a decision tree to fit}
\usage{easy_2var_data}
\format{Data frame containing 3 variables and 100 observations.}
\source{Author's creation.}
\keyword{datasets}