\name{grow.forest}
\alias{grow.forest}
\title{This is the title string for grow.forest}
\description{
	This is the title string for grow.forest
}
\usage{
grow.forest(formula, data, subset, weights, na.action,
    impurity.function = "gini", model = FALSE, x = FALSE, y = FALSE,
    min_node_obs, max_depth, 
    numsamps, numvars, numboots)
}
\arguments{
	\item{formula}{Needs documentation.}
	\item{data}{Needs documentation.}
	\item{subset}{Needs documentation.}
	\item{weights}{Needs documentation.}
	\item{na.action}{Needs documentation.}
	\item{impurity.function}{Needs documentation.}
	\item{model}{Needs documentation.}
	\item{x}{Needs documentation.}
	\item{y}{Needs documentation.}
	\item{min_node_obs}{Needs documentation.}
	\item{max_depth}{Needs documentation.}
	\item{numsamps}{Needs documentation.}
	\item{numvars}{Needs documentation.}
	\item{numboots}{Needs documentation.}
}