\name{forest-class}
\Rdversion{1.1}
\docType{class}
\alias{forest-class}


\title{Class \code{"forest"}}
\description{
A forest of decision tree classifiers used for ensemble prediction.
}
\section{Objects from the Class}{
Objects can be created by calls of the form \code{new("forest", ...)}.
%%  ~~ describe objects here ~~ 
}
\section{Slots}{
  \describe{
    \item{\code{n}:}{Object of class \code{"integer"} ~~ }
    \item{\code{p}:}{Object of class \code{"integer"} ~~ }
    \item{\code{min_node_obs}:}{Object of class \code{"integer"} ~~ }
    \item{\code{max_depth}:}{Object of class \code{"integer"} ~~ }
    \item{\code{numsamps}:}{Object of class \code{"integer"} ~~ }
    \item{\code{numvars}:}{Object of class \code{"integer"} ~~ }
    \item{\code{numboots}:}{Object of class \code{"integer"} ~~ }
    \item{\code{numnodes}:}{Object of class \code{"integer"} ~~ }
    \item{\code{flattened.nodes}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{model}:}{Object of class \code{"data.frame"} ~~ }
    \item{\code{x}:}{Object of class \code{"data.frameORvector"} ~~ }
    \item{\code{y}:}{Object of class \code{"vector"} ~~ }
    \item{\code{fmla}:}{Object of class \code{"formula"} ~~ }
  }
}

\keyword{classes}
