\name{low_high_earners}
\docType{data}
\alias{low_high_earners}
\title{Low earners and high earners}
\description{Dataset of low earners and high earners for classification.}
\usage{low_high_earners}
\format{Needs documentation.}
\source{Prepared from http://archive.ics.uci.edu/ml/datasets/Census-Income+\%28KDD\%29}
\keyword{datasets}