\name{low_high_earners}
\docType{data}
\alias{low_high_earners}
\title{Low earners and high earners, training dataset}
\description{Dataset of low earners and high earners for classification. \code{low_high_earners} is the training dataset and \code{low_high_earners_test} is the testing dataset.}
\usage{low_high_earners}
\format{Data frame containing 8 variables and 199,522 observations.}
\source{Prepared by keeping only the dependent variable, continuous variables, ordinal variables, and binary categorical variables from http://archive.ics.uci.edu/ml/datasets/Census-Income+\%28KDD\%29}
\keyword{datasets}