\name{predict.forest}
\alias{predict.forest}
\title{This is the title string for predict.forest}
\description{
	This is the title string for predict.forest
}
\usage{
predict.forest(object, newdata, \dots)
}
\arguments{
	\item{object}{Needs documentation.}
	\item{newdata}{Needs documentation.}
	\item{\dots}{Needs documentation.}
}