\name{predict.forest}
\alias{predict.forest}
\alias{predict,forest-method}
\title{Predict method for random decision forest classifier fits}
\description{
	Predict method for random decision forest classifier fits
}
\usage{
\S4method{predict}{forest}(object, newdata, \dots)
}
\arguments{
	\item{object}{Object of class inheriting from \code{"\link{forest-class}"}.}
	\item{newdata}{A data frame in which to look for variables with which to predict.}
	\item{\dots}{further arguments passed to or from other methods.}
}